requirements.txt
streamlit==1.28.1
requests==2.31.0
python-dotenv==1.0.0
elevenlabs==0.2.26
pydantic>=2.6.0
websockets==12.0
asyncio-mqtt==0.16.1
aiohttp==3.9.1 